Move IdempotencyRequestCache to `WalletWasabi` project.
