+++ WalletWasabi.Tests/UnitTests/WabiSabi/Client/ArenaClientTests.cs
@@ -6,8 +6,8 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using WalletWasabi.Backend.Controllers;
-using WalletWasabi.Backend.Controllers.WabiSabi;
 using WalletWasabi.BitcoinCore.Rpc;
+using WalletWasabi.Cache;
 using WalletWasabi.Crypto;
 using WalletWasabi.Crypto.Randomness;
 using WalletWasabi.Crypto.ZeroKnowledge;
