+++ WalletWasabi.Tests/UnitTests/WabiSabi/Client/BobClientTests.cs
@@ -4,8 +4,8 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using WalletWasabi.Backend.Controllers;
-using WalletWasabi.Backend.Controllers.WabiSabi;
 using WalletWasabi.Blockchain.TransactionOutputs;
+using WalletWasabi.Cache;
 using WalletWasabi.Crypto.Randomness;
 using WalletWasabi.Tests.Helpers;
 using WalletWasabi.WabiSabi;
